fix(dgaard/runtime): enforce allowed_networks ACL on incoming UDP packets
- Add is_ip_allowed() with CIDR matching via std::net (no new deps)
- Enforce ACL in both single-worker and multi-worker receive loops
- Packets from unlisted sources are silently dropped before task spawn
- Fix pre-existing clippy warnings across the workspace (cast, ptr_arg,
  derivable_impls, type_complexity, collapsible_if)

+++ dgaard-engine/src/config/parser.rs
@@ -659,7 +659,7 @@ fn parse_custom_flag(
     let bit = get_integer(table, "bit")?.ok_or_else(|| ConfigError::MissingKey {
         key: "bit".to_string(),
         span: parent_span,
-    })? as i64;
+    })?;
 
     if !(16..=31).contains(&bit) {
         return Err(ConfigError::InvalidValue {
